#13 feat: 거래 상세 스크롤 훅으로 분리
